requirements.txt
# /assessment_engine/requirements.txt

fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai